Fix various deprecation warnings

+++ easyunfold/pyvaspwfc/paw.py
@@ -166,10 +166,10 @@ class pawpotcar(object):
             for rr in dump[1:]:
                 reci, real = rr.split('Real Space Part')
                 qprojs.append(
-                    np.fromstring(reci, np.float, sep=' ')
+                    np.fromstring(reci, float, sep=' ')
                 )
                 rprojs.append(
-                    np.fromstring(real, np.float, sep=' ')
+                    np.fromstring(real, float, sep=' ')
                 )
 
         # the real space radial grid for the projector functions
